eliminacion de cuarto

+++ controller/cuartoController.js
@@ -3,6 +3,7 @@ const Cuarto = require("../models/cuartos");
 const Torre = require("../models/torres");
 const Piso = require("../models/pisos");
 const Beca = require("../models/becas");
+const Estudiantes = require("../models/estudiantes");
 
 const getCuarto = async () => {
   try {
@@ -57,9 +58,21 @@ const updateCuarto = async (id, numero_cuarto, capacidad_maxima, torreid) => {
 };
 const deleteCuarto = async (id) => {
   try {
+    // Verificar si hay estudiantes asociados al cuarto
+    const estudiantesAsociados = await Estudiantes.count({
+      where: { cuartoId: id },
+    });
+    if (estudiantesAsociados > 0) {
+      throw new AppError(
+        "No se puede eliminar el cuarto porque tiene estudiantes asociados.",
+        400
+      );
+    }
+
     const cuarto = await Cuarto.destroy({ where: { id } });
     return cuarto;
   } catch (error) {
+    console.error("Error al eliminar el cuarto:", error);
     throw error;
   }
 };
